requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
cbor2==5.9.0
celery==5.5.3
certifi==2025.6.15
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
chardet==7.4.0.post2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
cryptography==46.0.6
curl_cffi==0.11.4
cycler==0.12.1
daphne==4.2.1
defusedxml==0.7.1
dj-database-url==3.1.0
Django==5.2.4
django-appconf==1.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.4
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-spectacular==0.29.0
et_xmlfile==2.0.0
flake8==7.3.0
Flask==3.1.1
fonttools==4.58.5
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==20.1.0
h11==0.16.0
hiredis==3.3.1
httpcore==1.0.9
httplib2==0.30.0
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
kombu==5.5.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
msgpack==1.1.2
multitasking==0.0.11
narwhals==1.46.0
nltk==3.9.1
numpy==2.3.1
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
paramiko==4.0.0
peewee==3.18.1
pillow==11.0.0
pipreqs==0.4.13
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==26.0.0
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.37.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tabulate==0.9.0
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.42.0
vine==5.1.0
vulture==2.16
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.8.2
yarg==0.1.10
yfinance==0.2.64
zope.interface==8.2
